Users and authentication routes

+++ backend/src/services/posts.js
@@ -1,9 +1,10 @@
 import { Post } from '../db/models/posts.js'
+import { User } from '../db/models/user.js'
 //Imports the model of a post from the models/posts.js file
 
 //Creates a new post and saves it to the database
-export async function createPost({title, content, author, tags}) {
-    const post = new Post({title, author, content, tags})
+export async function createPost(userId, {title, content, tags}) {
+    const post = new Post({title, author: userId, content, tags})
     return await post.save()
 }
 async function listPosts(
@@ -15,8 +16,10 @@ async function listPosts(
 export async function listAllPosts(options){
     return await listPosts({}, options)
 }
-export async function listPostsByAuthor(author, options){
-    return await listPosts({author}, options)
+export async function listPostsByAuthor(authorUsername, options){
+    const user = await User.findOne({ username: authorUsername })
+    if (!user) return []
+    return await listPosts({author: user._id}, options)
 }
 export async function listPostsByTag(tags, options){
     return await listPosts({ tags }, options)
@@ -24,15 +27,15 @@ export async function listPostsByTag(tags, options){
 export async function getPostById(postId){
     return await Post.findById(postId)
 }
-export async function updatePost(postId, {title, author, content, tags}){
+export async function updatePost(userId, postId, {title, content, tags}){
     return await Post.findOneAndUpdate(
-        { _id: postId },
-        {$set: {title, author, content, tags}},
+        { _id: postId, author: userId },
+        {$set: {title, content, tags}},
         {new: true}
     )
 }
-export async function deletePost(postId){
-    return await Post.deleteOne({_id: postId})
+export async function deletePost(userId, postId){
+    return await Post.deleteOne({_id: postId, author: userId})
 }
